Widen num_rows * topK products in moe_permute_row_map for consistency
Per reviewer feedback in NVIDIA/TransformerEngine#2907, promote the
int * int multiplications in moe_permute_row_map and its launcher to
int64_t. These are not the overflow path this PR was originally
fixing (DeepSeek-V3 long-context hits row * num_cols, where num_cols
is the hidden dim ~ 7-8k), and num_rows * topK only crosses 2**31 at
unrealistic per-rank token counts (>= 268M at topK=8). The change is
purely defensive but keeps the index arithmetic in this kernel
consistent with the int64_t source_token / source_row / dest_row
widening already applied to moe_unpermute_kernel and moe_permute_kernel.

+++ transformer_engine/common/permutation/permutation.cu
@@ -19,7 +19,7 @@ static __global__ void moe_permute_row_map(const int *sorted_row_id, int *row_id
   const int tid = threadIdx.x;
   const int idx = bid * blockDim.x + tid;
 
-  if (idx >= num_rows * topK) return;
+  if (idx >= static_cast<int64_t>(num_rows) * topK) return;
 
   int source_row = sorted_row_id[idx];
   int source_token_id = source_row / topK;
@@ -27,10 +27,10 @@ static __global__ void moe_permute_row_map(const int *sorted_row_id, int *row_id
 
   if (idx >= num_out_tokens) {
     // Set the indices of dropped tokens to -1
-    row_id_map[source_topK_id * num_rows + source_token_id] = -1;
+    row_id_map[static_cast<int64_t>(source_topK_id) * num_rows + source_token_id] = -1;
   } else {
     // Create a row id map for subsequent unpermute operation
-    row_id_map[source_topK_id * num_rows + source_token_id] = idx;
+    row_id_map[static_cast<int64_t>(source_topK_id) * num_rows + source_token_id] = idx;
   }
 }
 
@@ -239,7 +239,7 @@ void nvte_permute_launcher(const T *input, T *output, const int *sorted_row_id, 
     // moe_permute_fwd
 
     int threads = 64;
-    int blocks = (num_rows * topK + threads - 1) / threads;
+    int blocks = (static_cast<int64_t>(num_rows) * topK + threads - 1) / threads;
 
     moe_permute_row_map<<<blocks, threads, 0, stream>>>(sorted_row_id, row_id_map, num_rows, topK,
                                                         num_out_tokens);
